is: Move store field mask utilities to separate files
