+++ pkg/identityserver/store/field_mask_test.go
@@ -1,0 +1,46 @@
+// Copyright © 2022 The Things Network Foundation, The Things Industries B.V.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package store
+
+import (
+	"testing"
+
+	"go.thethings.network/lorawan-stack/v3/pkg/util/test"
+	"go.thethings.network/lorawan-stack/v3/pkg/util/test/assertions/should"
+)
+
+func TestFieldMask_Contains(t *testing.T) {
+	t.Parallel()
+	a, _ := test.New(t)
+
+	a.So(FieldMask{}.Contains(""), should.BeFalse)
+	a.So(FieldMask{"a", "b", "c"}.Contains("a"), should.BeTrue)
+}
+
+func TestFieldMask_TopLevel(t *testing.T) {
+	t.Parallel()
+	a, _ := test.New(t)
+
+	a.So(FieldMask{}.TopLevel(), should.BeEmpty)
+	a.So(FieldMask{"a.foo", "a.bar", "b", "c"}.TopLevel(), should.Resemble, FieldMask{"a", "b", "c"})
+}
+
+func TestFieldMask_TrimPrefix(t *testing.T) {
+	t.Parallel()
+	a, _ := test.New(t)
+
+	a.So(FieldMask{}.TrimPrefix("a"), should.BeEmpty)
+	a.So(FieldMask{"a.foo", "a.bar", "b", "c"}.TrimPrefix("a"), should.Resemble, FieldMask{"foo", "bar"})
+}
